# internal/api/models.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package api

// GetFirstChallengeResponse Response containing the first challenge details
type GetFirstChallengeResponse struct {
	// Challenge The challenge string that needs to be solved
	Challenge string `json:"challenge"`

	// Difficulty Difficulty level of the challenge
	Difficulty int32 `json:"difficulty"`

	// Token Temporary token for challenge verification
	Token string `json:"token"`

	// UserId Temporary user ID for the challenge session
	UserId string `json:"userId"`
}

// RegisterUserRequest Request to register a new user
type RegisterUserRequest struct {
	// Challenge The solved challenge string
	Challenge string `json:"challenge"`

	// Difficulty Difficulty level of the solved challenge
	Difficulty int32 `json:"difficulty"`

	// Nickname User's chosen display name
	Nickname string `json:"nickname"`

	// PublicKey User's public key for secure communication
	PublicKey string `json:"publicKey"`

	// Solution The solution to the challenge
	Solution string   `json:"solution"`
	Tags     []string `json:"tags"`

	// UserId Verified user ID from challenge solution
	UserId string `json:"userId"`
}

// RegisterUserResponse Response after user registration attempt
type RegisterUserResponse struct {
	// Message Additional information about the registration result
	Message string `json:"message"`

	// Success Whether the registration was successful
	Success bool `json:"success"`

	// UserId Registered user ID
	UserId string `json:"userId"`
}

// SolveFirstChallengeRequest Request to solve the first challenge
type SolveFirstChallengeRequest struct {
	// Challenge The challenge string that was received
	Challenge string `json:"challenge"`

	// Difficulty Difficulty level of the challenge
	Difficulty int32 `json:"difficulty"`

	// Solution The solution to the challenge
	Solution string `json:"solution"`

	// Token Temporary token received from the challenge
	Token string `json:"token"`

	// UserId Temporary user ID received from the challenge
	UserId string `json:"userId"`
}

// SolveFirstChallengeResponse Response after successful challenge solution
type SolveFirstChallengeResponse struct {
	// Challenge The solved challenge string
	Challenge string `json:"challenge"`

	// Difficulty Difficulty level of the solved challenge
	Difficulty int32 `json:"difficulty"`

	// UserId Verified user ID for registration
	UserId string `json:"userId"`
}

// SolveFirstChallengeJSONRequestBody defines body for SolveFirstChallenge for application/json ContentType.
type SolveFirstChallengeJSONRequestBody = SolveFirstChallengeRequest

// RegisterUserJSONRequestBody defines body for RegisterUser for application/json ContentType.
type RegisterUserJSONRequestBody = RegisterUserRequest